+++ belokurskaya.sofia/P3/transform_string.hpp
@@ -1,0 +1,11 @@
+#ifndef TRANSFORM_STRING_HPP
+#define TRANSFORM_STRING_HPP
+
+#include <cstring>
+#include <cctype>
+
+void transformString(const char * input, char * result, char * not_input_chars);
+
+void toLowerCase(char * input);
+
+#endif
